+++ cmd/mgrs/to_wgs84.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"strings"
+	"math"
 
 	"github.com/brundtoe/go-geografi/pkg/mgrs"
 	"github.com/brundtoe/go-geografi/pkg/utils"
@@ -38,29 +38,23 @@ func main() {
 		}
 		// The first line contains field names
 		if record[1] != "City" {
-			transformMGRS(record)
+			transformWgs84(record)
 		}
 	}
 
 }
 
-func transformMGRS(record []string) {
+func transformWgs84(record []string) {
 	location := mgrs.City{}
 	location.BuildCity(record)
-	// de to funktioner kalder undervejs LL.ToUTM()
-	milgrid := location.CityToMgrs()
-	usng := location.CityToUsng()
-	fmt.Printf("%-18s %s %s", location.Name, milgrid, usng)
+	ll, _ := location.Utm.ToLL()
+	fmt.Printf("%20s %-18s", ll.String(), location.Name)
 
-	east := string(usng[7:12])
-	north := string(usng[13:18])
-	mgrsEast := string(milgrid[5:10])
-	mgrsNorth := string(milgrid[10:15])
-	if strings.Compare(east, record[11]) != 0 || strings.Compare(mgrsEast, record[11]) != 0 {
-		fmt.Print(" East")
+	if math.Abs(ll.Lon-location.Geoloc.Lon) > 0.000001 {
+		fmt.Printf("%10.6f", location.Geoloc.Lon)
 	}
-	if strings.Compare(north, record[12]) != 0 || strings.Compare(mgrsNorth, record[12]) != 0 {
-		fmt.Print(" North")
+	if math.Abs(ll.Lat-location.Geoloc.Lat) > 0.000001 {
+		fmt.Printf("Lat %10.6f", location.Geoloc.Lat)
 	}
-	fmt.Println("")
+	fmt.Println()
 }